+++ src/main/java/com/imsweb/mph/mpgroups/Mp2010HematopoieticGroup.java
@@ -5,6 +5,7 @@ package com.imsweb.mph.mpgroups;
 
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 
 import com.imsweb.mph.MphComputeOptions;
@@ -13,6 +14,7 @@ import com.imsweb.mph.MphGroup;
 import com.imsweb.mph.MphInput;
 import com.imsweb.mph.MphRule;
 import com.imsweb.mph.MphUtils;
+import com.imsweb.mph.MphUtils.MpResult;
 import com.imsweb.mph.internal.TempRuleResult;
 
 public class Mp2010HematopoieticGroup extends MphGroup {
@@ -37,17 +39,25 @@ public class Mp2010HematopoieticGroup extends MphGroup {
             @Override
             public TempRuleResult apply(MphInput i1, MphInput i2, MphComputeOptions options) {
                 TempRuleResult result = new TempRuleResult();
-                //TODO
+                String h1 = i1.getHistology(), h2 = i2.getHistology(), s1 = i1.getPrimarySite(), s2 = i2.getPrimarySite();
+                if (h1.equals(h2)) {
+                    boolean nodalAndExtraNodalMalt = MphConstants.MALT.equals(i1.getMorphology()) && MphConstants.MALT.equals(i2.getMorphology()) && ((s1.startsWith("C77") && !s2.startsWith("C77"))
+                            || (s2.startsWith("C77") && !s1.startsWith("C77")));
+                    result.setFinalResult(nodalAndExtraNodalMalt ? MpResult.QUESTIONABLE : MpResult.SINGLE_PRIMARY);
+                }
                 return result;
             }
         };
-        rule.setReason("Abstract a single primary when there is a single histology");
+        rule.setReason(
+                "Abstract a single primary when there is a single histology. Exception: manually review when a nodal MALT (C770-779, 9699/3) occurs before or after an extranodal MALT (all other sites, 9699/3). These are two distinct lymphomas that have the same histology code.");
         rule.getNotes().add("Bilateral involvement of lymph nodes and/or organs with a single histology is a single primary.");
         rule.getNotes().add("Recurrence of the same histology is always the same primary (timing is not relevant).");
         rule.getNotes().add("A single histology is diagnosed by the definitive diagnostic method as defined in the Heme DB. For example, the patient had several provisional diagnoses "
                 + "but the definitive diagnostic method identified a single histology. Abstract as a single primary.");
         rule.getExamples().add("The diagnosis is multiple myeloma (9732/3). Abstract as a single primary.");
         rule.getExamples().add("Right and left breast both involved with diffuse large B-cell lymphoma (9680/3). Abstract as a single primary.");
+        rule.getExamples().add(
+                "Marginal zone lymphoma (MALT) of right inguinal node (C774) diagnosed in 2013. Stage I with no recurrence. In March 2018, diagnosed with Stage III ocular marginal zone lymphoma. Abstract a new primary.");
         _rules.add(rule);
 
         // M3
@@ -74,9 +84,9 @@ public class Mp2010HematopoieticGroup extends MphGroup {
                 return result;
             }
         };
-        rule.setReason("Abstract a single primary when a sarcoma is diagnosed simultaneously or after a leukemia of the same lineage, " +
-                "Mast cell sarcoma (9740/3) diagnosed simultaneously with or after mast cell leukemia (9742/3), " +
-                "Myeloid sarcoma (9930/3) diagnosed simultaneously with or after acute myeloid leukemia (9861/3) or another leukemia of the myeloid lineage (9840/3, 9865/3-9867/3, 9869/3-9874/3, 9891/3, 9895/3-9898/3, 9910/3, 9911/3 and 9931/3)"
+        rule.setReason("Abstract a single primary when a sarcoma is diagnosed simultaneously or after a leukemia of the same lineage,\n" +
+                " - Mast cell sarcoma (9740/3) diagnosed simultaneously with or after mast cell leukemia (9742/3),\n" +
+                " - Myeloid sarcoma (9930/3) diagnosed simultaneously with or after acute myeloid leukemia (9861/3) or another leukemia of the myeloid lineage (9840/3, 9865/3-9867/3, 9869/3-9874/3, 9891/3, 9895/3-9898/3, 9910/3, 9911/3 and 9931/3)\n"
                 + "Exception: Chronic myeloid leukemia (CML) codes: 9863/3, 9875/3, 9876/3 are not classified as leukemias of the same lineage as myeloid sarcoma");
         rule.getNotes().add("These sarcomas are solid manifestations of the associated leukemias. For example, when acute myeloid leukemia and myeloid sarcoma are diagnosed "
                 + "simultaneously, the myeloid sarcoma is the result of myeloid cells migrating from the bone marrow or blood into tissue. It is part of the disease process "
@@ -192,26 +202,12 @@ public class Mp2010HematopoieticGroup extends MphGroup {
             @Override
             public TempRuleResult apply(MphInput i1, MphInput i2, MphComputeOptions options) {
                 TempRuleResult result = new TempRuleResult();
-                String hist1 = i1.getHistology(), hist2 = i2.getHistology();
-                String morph1 = hist1 + "/" + i1.getBehavior(), morph2 = hist2 + "/" + i2.getBehavior();
-                int latestDx = GroupUtility.compareDxDate(i1, i2);
-                int year1 = Integer.valueOf(i1.getDateOfDiagnosisYear()), year2 = Integer.valueOf(i2.getDateOfDiagnosisYear());
-                if (!MphConstants.HEMATOPOIETIC_NOS_HISTOLOGIES.containsAll(Arrays.asList(hist1, hist2)) && (MphConstants.HEMATOPOIETIC_NOS_HISTOLOGIES.contains(hist1)
-                        || MphConstants.HEMATOPOIETIC_NOS_HISTOLOGIES.contains(hist2)) && MphUtils.getInstance().getHematoDbUtilsProvider().isSamePrimary(morph1, morph2, year1, year2)
-                        && MphConstants.COMPARE_DX_EQUAL != latestDx) {
-                    if (MphConstants.COMPARE_DX_UNKNOWN == latestDx) {
-                        result.setPotentialResult(MphUtils.MpResult.SINGLE_PRIMARY);
-                        result.setMessageUnknownDiagnosisDate(this.getStep(), this.getGroupId());
-                    }
-                    else if ((MphConstants.COMPARE_DX_FIRST_LATEST == latestDx && MphConstants.HEMATOPOIETIC_NOS_HISTOLOGIES.contains(hist2)) || (MphConstants.COMPARE_DX_SECOND_LATEST == latestDx
-                            && MphConstants.HEMATOPOIETIC_NOS_HISTOLOGIES.contains(hist1)))
-                        result.setFinalResult(MphUtils.MpResult.SINGLE_PRIMARY);
-                }
+                //This rule is skipped by the automated process
                 return result;
             }
         };
         rule.setReason("Abstract as a single primary when a more specific histology is diagnosed after an NOS ONLY when the Heme DB Multiple Primaries Calculator "
-                + "confirms that the NOS and the more specific histology are the same primary.");
+                + "confirms that the NOS and the more specific histology are the same primary. Rule M7 skipped by automated rules. The Multiple Primaries Calculator will be checked for all applicable cases at Rule M15.");
         rule.getNotes().add("The more specific histology confirmation does not have to occur in the same anatomic location. ");
         rule.getNotes().add("There are no time restrictions on these diagnoses; the interval between the NOS and the more specific histology does not affect this rule.");
         rule.getNotes().add("The Heme DB Multiple Primaries Calculator will identify these histologies as a single primary");
@@ -231,7 +227,21 @@ public class Mp2010HematopoieticGroup extends MphGroup {
             @Override
             public TempRuleResult apply(MphInput i1, MphInput i2, MphComputeOptions options) {
                 TempRuleResult result = new TempRuleResult();
-                //TODO
+                String morph1 = i1.getHistology() + "/" + i1.getBehavior();
+                String morph2 = i2.getHistology() + "/" + i2.getBehavior();
+                int year1 = Integer.parseInt(i1.getDateOfDiagnosisYear()), year2 = Integer.parseInt(i2.getDateOfDiagnosisYear());
+                //If one disease can not be converted to another, no need to check other criteria
+                if (isTransformation(morph1, morph2, year1, year2)) {
+                    int daysApart = GroupUtility.verifyDaysApart(i1, i2, 21);
+                    //TODO, implement this correctly if we find a way to determine whether there is a documentation for biopsy or not programmatically
+                    //For now return manual review if the cases are diagnosed simultaneously or within 21 days
+                    if (daysApart != MphConstants.DATE_VERIFY_APART) {
+                        result.setFinalResult(MpResult.QUESTIONABLE);
+                        result.setMessage(
+                                "Needs a manual review. Abstract as a single primary and code the acute neoplasm when both a chronic and an acute neoplasm are diagnosed simultaneously or within 21 days AND there is documentation of only one positive biopsy (bone marrow biopsy, lymph node biopsy, or tissue biopsy).");
+                    }
+                }
+
                 return result;
             }
         };
@@ -250,7 +260,21 @@ public class Mp2010HematopoieticGroup extends MphGroup {
             @Override
             public TempRuleResult apply(MphInput i1, MphInput i2, MphComputeOptions options) {
                 TempRuleResult result = new TempRuleResult();
-                //TODO
+                String morph1 = i1.getHistology() + "/" + i1.getBehavior();
+                String morph2 = i2.getHistology() + "/" + i2.getBehavior();
+                int year1 = Integer.parseInt(i1.getDateOfDiagnosisYear()), year2 = Integer.parseInt(i2.getDateOfDiagnosisYear());
+                //If one disease can not be converted to another, no need to check other criteria
+                if (isTransformation(morph1, morph2, year1, year2)) {
+                    int daysApart = GroupUtility.verifyDaysApart(i1, i2, 21);
+                    //TODO, implement this correctly if we find a way to determine whether there is a documentation for biopsy or not programmatically
+                    //For now return manual review if the cases are diagnosed simultaneously or within 21 days
+                    if (daysApart != MphConstants.DATE_VERIFY_APART) {
+                        result.setFinalResult(MpResult.QUESTIONABLE);
+                        result.setMessage(
+                                "Needs a manual review. Abstract a single primary and code the later diagnosis when both a chronic and an acute neoplasm are diagnosed simultaneously or within 21 days "
+                                        + "AND there is no available documentation on biopsy (bone marrow biopsy, lymph node biopsy, or tissue biopsy.) The later diagnosis could be either the chronic or the acute neoplasm.");
+                    }
+                }
                 return result;
             }
         };
@@ -270,9 +294,14 @@ public class Mp2010HematopoieticGroup extends MphGroup {
                 TempRuleResult result = new TempRuleResult();
                 String morph1 = i1.getHistology() + "/" + i1.getBehavior();
                 String morph2 = i2.getHistology() + "/" + i2.getBehavior();
-                int year1 = Integer.valueOf(i1.getDateOfDiagnosisYear()), year2 = Integer.valueOf(i2.getDateOfDiagnosisYear());
+                int year1 = Integer.parseInt(i1.getDateOfDiagnosisYear()), year2 = Integer.parseInt(i2.getDateOfDiagnosisYear());
                 //If one disease can not be converted to another, no need to check other criteria
                 if (isTransformation(morph1, morph2, year1, year2)) {
+                    if (GroupUtility.differentCategory(morph1, morph2, Collections.singletonList("9732/3"), Arrays.asList("9731/3", "9734/3"))) {
+                        result.setMessage("For plasmacytoma (9731, 9734) and plasma cell myeloma (9732): This rule would only apply if the initial workup was completed and a single plasmacytoma was diagnosed. If plasma cell myeloma is diagnosed after the initial workup and treatment, then this rule would be applicable and the multiple myeloma would be a second primary.");
+                        result.setFinalResult(MpResult.QUESTIONABLE);
+                        return result;
+                    }
                     int latestDx = GroupUtility.compareDxDate(i1, i2);
                     int daysApart = GroupUtility.verifyDaysApart(i1, i2, 21);
                     if (MphConstants.DATE_VERIFY_UNKNOWN == daysApart || MphConstants.COMPARE_DX_UNKNOWN == latestDx) {
@@ -289,6 +318,7 @@ public class Mp2010HematopoieticGroup extends MphGroup {
         };
         rule.setReason("Abstract as multiple primaries when a neoplasm is originally diagnosed as a chronic neoplasm AND there is a second diagnosis of an acute "
                 + "neoplasm more than 21 days after the chronic diagnosis.");
+        rule.getNotes().add("The presence of multiple plasmacytomas is diagnostic of multiple myeloma, see Hematopoietic Database, 9732/3.");
         rule.getNotes().add("This is a change from the pre-2010 rules. Use the Heme DB Multiple Primaries Calculator to determine multiple primaries when a transformation "
                 + "from a chronic to an acute neoplasm occurs.");
         rule.getNotes().add("Transformations to (acute neoplasms) and Transformations from (chronic neoplasms) are defined for each applicable histology in the database.");
@@ -302,7 +332,21 @@ public class Mp2010HematopoieticGroup extends MphGroup {
             @Override
             public TempRuleResult apply(MphInput i1, MphInput i2, MphComputeOptions options) {
                 TempRuleResult result = new TempRuleResult();
-                //TODO
+                String morph1 = i1.getHistology() + "/" + i1.getBehavior();
+                String morph2 = i2.getHistology() + "/" + i2.getBehavior();
+                int year1 = Integer.parseInt(i1.getDateOfDiagnosisYear()), year2 = Integer.parseInt(i2.getDateOfDiagnosisYear());
+                //If one disease can not be converted to another, no need to check other criteria
+                if (isTransformation(morph1, morph2, year1, year2)) {
+                    int daysApart = GroupUtility.verifyDaysApart(i1, i2, 21);
+                    //TODO, implement this correctly if we find a way to determine whether there is a documentation for biopsy or not programmatically
+                    //For now return manual review if the cases are diagnosed simultaneously or within 21 days
+                    if (daysApart != MphConstants.DATE_VERIFY_APART) {
+                        result.setFinalResult(MpResult.QUESTIONABLE);
+                        result.setMessage(
+                                "Needs a manual review. Abstract as multiple primaries when both a chronic and an acute neoplasm are diagnosed simultaneously or"
+                                        + " within 21 days AND there is documentation of two bone marrow examinations, lymph node biopsies, or tissue biopsies: one confirming the chronic neoplasm and another confirming the acute neoplasm.");
+                    }
+                }
                 return result;
             }
         };
@@ -322,19 +366,11 @@ public class Mp2010HematopoieticGroup extends MphGroup {
                 TempRuleResult result = new TempRuleResult();
                 String morph1 = i1.getHistology() + "/" + i1.getBehavior();
                 String morph2 = i2.getHistology() + "/" + i2.getBehavior();
-                int year1 = Integer.valueOf(i1.getDateOfDiagnosisYear()), year2 = Integer.valueOf(i2.getDateOfDiagnosisYear());
+                int year1 = Integer.parseInt(i1.getDateOfDiagnosisYear()), year2 = Integer.parseInt(i2.getDateOfDiagnosisYear());
                 //If one disease can not be converted to another, no need to check other criteria
                 if (isTransformation(morph1, morph2, year1, year2)) {
-                    int latestDx = GroupUtility.compareDxDate(i1, i2);
-                    if (MphConstants.COMPARE_DX_UNKNOWN == latestDx) {
-                        result.setPotentialResult(MphUtils.MpResult.SINGLE_PRIMARY);
-                        result.setMessageUnknownDiagnosisDate(this.getStep(), this.getGroupId());
-                    }
-                    else if (latestDx > 0 && isAcuteToChronicTransformation(MphConstants.COMPARE_DX_FIRST_LATEST == latestDx ? morph2 : morph1,
-                            MphConstants.COMPARE_DX_FIRST_LATEST == latestDx ? morph1 : morph2, MphConstants.COMPARE_DX_FIRST_LATEST == latestDx ? year2 : year1,
-                            MphConstants.COMPARE_DX_FIRST_LATEST == latestDx ? year1 : year2) && !MphConstants.TREATMENT_GIVEN.equals(
-                            MphConstants.COMPARE_DX_FIRST_LATEST == latestDx ? i2.getTxStatus() : i1.getTxStatus()))
-                        result.setFinalResult(MphUtils.MpResult.SINGLE_PRIMARY);
+                    result.setFinalResult(MpResult.QUESTIONABLE);
+                    result.setMessage("Manual review is required to check whether there is confirmation of treatment or not.");
                 }
                 return result;
             }
@@ -358,19 +394,11 @@ public class Mp2010HematopoieticGroup extends MphGroup {
                 String morph1 = i1.getHistology() + "/" + i1.getBehavior();
                 String morph2 = i2.getHistology() + "/" + i2.getBehavior();
 
-                int year1 = Integer.valueOf(i1.getDateOfDiagnosisYear()), year2 = Integer.valueOf(i2.getDateOfDiagnosisYear());
+                int year1 = Integer.parseInt(i1.getDateOfDiagnosisYear()), year2 = Integer.parseInt(i2.getDateOfDiagnosisYear());
                 //If one disease can not be converted to another, no need to check other criteria
                 if (isTransformation(morph1, morph2, year1, year2)) {
-                    int latestDx = GroupUtility.compareDxDate(i1, i2);
-                    if (MphConstants.COMPARE_DX_UNKNOWN == latestDx) {
-                        result.setPotentialResult(MphUtils.MpResult.MULTIPLE_PRIMARIES);
-                        result.setMessageUnknownDiagnosisDate(this.getStep(), this.getGroupId());
-                    }
-                    else if (latestDx > 0 && isAcuteToChronicTransformation(MphConstants.COMPARE_DX_FIRST_LATEST == latestDx ? morph2 : morph1,
-                            MphConstants.COMPARE_DX_FIRST_LATEST == latestDx ? morph1 : morph2, MphConstants.COMPARE_DX_FIRST_LATEST == latestDx ? year2 : year1,
-                            MphConstants.COMPARE_DX_FIRST_LATEST == latestDx ? year1 : year2) && MphConstants.TREATMENT_GIVEN.equals(
-                            MphConstants.COMPARE_DX_FIRST_LATEST == latestDx ? i2.getTxStatus() : i1.getTxStatus()))
-                        result.setFinalResult(MphUtils.MpResult.MULTIPLE_PRIMARIES);
+                    result.setFinalResult(MpResult.QUESTIONABLE);
+                    result.setMessage("Manual review is required to check whether there is confirmation of treatment or not.");
                 }
                 return result;
             }
@@ -425,13 +453,13 @@ public class Mp2010HematopoieticGroup extends MphGroup {
             public TempRuleResult apply(MphInput i1, MphInput i2, MphComputeOptions options) {
                 TempRuleResult result = new TempRuleResult();
                 String morph1 = i1.getHistology() + "/" + i1.getBehavior(), morph2 = i2.getHistology() + "/" + i2.getBehavior();
-                int year1 = Integer.valueOf(i1.getDateOfDiagnosisYear()), year2 = Integer.valueOf(i2.getDateOfDiagnosisYear());
+                int year1 = Integer.parseInt(i1.getDateOfDiagnosisYear()), year2 = Integer.parseInt(i2.getDateOfDiagnosisYear());
                 result.setFinalResult(
                         MphUtils.getInstance().getHematoDbUtilsProvider().isSamePrimary(morph1, morph2, year1, year2) ? MphUtils.MpResult.SINGLE_PRIMARY : MphUtils.MpResult.MULTIPLE_PRIMARIES);
                 return result;
             }
         };
-        rule.setReason("Use the Heme DB Multiple Primaries Calculator to determine the number of primaries for all cases that do not meet the criteria of M1-M14.");
+        rule.setReason("Use the Heme DB Multiple Primaries Calculator to determine the number of primaries for all cases that do not meet the criteria of M1-M14 (Rule M7 executed by automated rules at this step).");
         rule.getExamples().add("Polycythemia vera (PV) diagnosed in 2001, receiving anagrelide. Increasing leukocytosis seen, bone marrow biopsy done in 2013 showing primary "
                 + "myelofibrosis (PMF) with myeloid metaplasia. No rule in M1-M14 applies. Abstract multiple primaries because the Multiple Primaries Calculator shows that PV "
                 + "(9950/3) and PMF (9961/3) are separate primaries.");