+++ src/test/java/com/imsweb/mph/MphUtilsTest.java
@@ -1864,7 +1864,32 @@ public class MphUtilsTest {
 
         //M1 TODO
 
-        //M2 TODO
+        //M2
+        i1.setPrimarySite("C779");
+        i2.setPrimarySite("C189");
+        i1.setHistologyIcdO3("9740");
+        i2.setHistologyIcdO3("9740");
+        i1.setBehaviorIcdO3("3");
+        i2.setBehaviorIcdO3("3");
+        i1.setDateOfDiagnosisYear("2015");
+        i2.setDateOfDiagnosisYear("2010");
+        output = _utils.computePrimaries(i1, i2);
+        Assert.assertEquals(MphUtils.MpResult.SINGLE_PRIMARY, output.getResult());
+        Assert.assertEquals(2, output.getAppliedRules().size());
+        Assert.assertTrue(output.getReason().contains("single histology"));
+        //Exception
+        i1.setHistologyIcdO3("9699");
+        i2.setHistologyIcdO3("9699");
+        output = _utils.computePrimaries(i1, i2);
+        Assert.assertEquals(MpResult.QUESTIONABLE, output.getResult());
+        Assert.assertEquals(2, output.getAppliedRules().size());
+        Assert.assertTrue(output.getReason().contains("single histology"));
+        i1.setPrimarySite("C779");
+        i2.setPrimarySite("C770");
+        output = _utils.computePrimaries(i1, i2);
+        Assert.assertEquals(MphUtils.MpResult.SINGLE_PRIMARY, output.getResult());
+        Assert.assertEquals(2, output.getAppliedRules().size());
+        Assert.assertTrue(output.getReason().contains("single histology"));
 
         //M3 Abstract a single primary* when a sarcoma is diagnosed simultaneously or after a leukemia of the same lineage
         i1.setPrimarySite("C408");
@@ -1910,7 +1935,7 @@ public class MphUtilsTest {
         i1.setPrimarySite("C771");
         i2.setPrimarySite("C771");
         i1.setHistologyIcdO3("9590");
-        i2.setHistologyIcdO3("9729");
+        i2.setHistologyIcdO3("9738");
         i1.setBehaviorIcdO3("3");
         i2.setBehaviorIcdO3("3");
         i1.setDateOfDiagnosisYear("2010");
@@ -1923,19 +1948,6 @@ public class MphUtilsTest {
         Assert.assertEquals(MphUtils.MpResult.SINGLE_PRIMARY, output.getResult());
         Assert.assertEquals(4, output.getAppliedRules().size());
         Assert.assertTrue(output.getReason().contains("non-Hodgkin"));
-        i2.setDateOfDiagnosisDay(null);
-        output = _utils.computePrimaries(i1, i2);
-        Assert.assertEquals(MphUtils.MpResult.SINGLE_PRIMARY, output.getResult());
-        Assert.assertEquals(15, output.getAppliedRules().size());
-        i1.setPrimarySite("C771");
-        i2.setPrimarySite("C772"); //not same location
-        output = _utils.computePrimaries(i1, i2);
-        Assert.assertNotEquals(4, output.getAppliedRules().size());
-        i1.setPrimarySite("C181");
-        i2.setPrimarySite("C182"); //same location
-        output = _utils.computePrimaries(i1, i2);
-        Assert.assertEquals(MphUtils.MpResult.SINGLE_PRIMARY, output.getResult());
-        Assert.assertEquals(15, output.getAppliedRules().size());
         i2.setDateOfDiagnosisYear("2011"); //not simultaneous
         output = _utils.computePrimaries(i1, i2);
         Assert.assertNotEquals(4, output.getAppliedRules().size());
@@ -1947,7 +1959,7 @@ public class MphUtilsTest {
         i1.setPrimarySite("C771");
         i2.setPrimarySite("C771");
         i1.setHistologyIcdO3("9653");
-        i2.setHistologyIcdO3("9700");
+        i2.setHistologyIcdO3("9738");
         i1.setBehaviorIcdO3("3");
         i2.setBehaviorIcdO3("3");
         i1.setDateOfDiagnosisYear("2010");
@@ -1984,42 +1996,8 @@ public class MphUtilsTest {
 
         //M7 Abstract as a single primary* when a more specific histology is diagnosed after an NOS ONLY when the Heme DB Multiple Primaries Calculator
         //confirms that the NOS and the more specific histology are the same primary.
-        i1 = new MphInput();
-        i2 = new MphInput();
-        i1.setPrimarySite("C001");
-        i2.setPrimarySite("C772");
-        i1.setHistologyIcdO3("9591");
-        i2.setHistologyIcdO3("9861"); //Both Nos histologies
-        i1.setBehaviorIcdO3("3");
-        i2.setBehaviorIcdO3("3");
-        i1.setDateOfDiagnosisYear("2010");
-        i2.setDateOfDiagnosisYear("2010");
-        output = _utils.computePrimaries(i1, i2);
-        Assert.assertNotEquals(7, output.getAppliedRules().size());
-        i1.setHistologyIcdO3("9671");
-        i2.setHistologyIcdO3("9672"); //Both Specific histologies
-        output = _utils.computePrimaries(i1, i2);
-        Assert.assertNotEquals(7, output.getAppliedRules().size());
-        i1.setHistologyIcdO3("9702");
-        i2.setHistologyIcdO3("9706"); //Nos vs specific, but not same primaries according to Hemato DB
-        output = _utils.computePrimaries(i1, i2);
-        Assert.assertNotEquals(7, output.getAppliedRules().size());
-        i1.setHistologyIcdO3("9702");
-        i2.setHistologyIcdO3("9705"); //Nos vs specific, same primaries according to Hemato DB
-        output = _utils.computePrimaries(i1, i2);
-        //Questionable at M7 with potential single and ended up as single at M15 -- SINGLE
-        Assert.assertEquals(15, output.getAppliedRules().size());
-        Assert.assertEquals(MphUtils.MpResult.SINGLE_PRIMARY, output.getResult());
-        Assert.assertEquals("M15", output.getStep());
-        i1.setDateOfDiagnosisYear("2011");
-        i2.setDateOfDiagnosisYear("2010"); //More specific was before Nos, not M7
-        output = _utils.computePrimaries(i1, i2);
-        Assert.assertNotEquals(7, output.getAppliedRules().size());
-        i1.setDateOfDiagnosisYear("2009");
-        i2.setDateOfDiagnosisYear("2010"); //More specific is after Nos, single at M7
-        output = _utils.computePrimaries(i1, i2);
-        Assert.assertEquals(7, output.getAppliedRules().size());
-        Assert.assertEquals(MphUtils.MpResult.SINGLE_PRIMARY, output.getResult());
+        //This is Skipped on the automated process
+
 
         //M8 TODO
 
@@ -2042,7 +2020,7 @@ public class MphUtilsTest {
         i1.setHistologyIcdO3("9875");
         i2.setHistologyIcdO3("9867");//9875 (chronic) transforms to 9867(acute)
         output = _utils.computePrimaries(i1, i2);
-        Assert.assertEquals(10, output.getAppliedRules().size());
+        Assert.assertEquals(8, output.getAppliedRules().size());
         Assert.assertEquals(MphUtils.MpResult.QUESTIONABLE, output.getResult());
         i1.setDateOfDiagnosisYear("2015");
         i2.setDateOfDiagnosisYear("2010"); // acute was diagnosed before chronic
@@ -2053,11 +2031,18 @@ public class MphUtilsTest {
         output = _utils.computePrimaries(i1, i2);
         Assert.assertEquals(10, output.getAppliedRules().size());
         Assert.assertEquals(MphUtils.MpResult.MULTIPLE_PRIMARIES, output.getResult());
+        //Exception for plasmacytoma (9731, 9734) and plasma cell myeloma (9732)
+        i1.setHistologyIcdO3("9731");
+        i2.setHistologyIcdO3("9732");
+        output = _utils.computePrimaries(i1, i2);
+        Assert.assertEquals(10, output.getAppliedRules().size());
+        Assert.assertEquals(MpResult.QUESTIONABLE, output.getResult());
 
         //M11 TODO
 
         //M12 Abstract a single primary* when a neoplasm is originally diagnosed as acute AND reverts to a chronic neoplasm AND there is no confirmation
         //available that the patient has been treated for the acute neoplasm.
+        //M12 and M13 are returning manual review
         i1 = new MphInput();
         i2 = new MphInput();
         i1.setPrimarySite("C001");
@@ -2070,17 +2055,13 @@ public class MphUtilsTest {
         i2.setDateOfDiagnosisYear("2015"); //acute before chrnoic
         output = _utils.computePrimaries(i1, i2);
         Assert.assertEquals(12, output.getAppliedRules().size());
-        Assert.assertEquals(MphUtils.MpResult.SINGLE_PRIMARY, output.getResult());
-        i2.setTxStatus("1"); //treatment was done for chronic, it doesn't matter
-        output = _utils.computePrimaries(i1, i2);
-        Assert.assertEquals(12, output.getAppliedRules().size());
-        Assert.assertEquals(MphUtils.MpResult.SINGLE_PRIMARY, output.getResult());
+        Assert.assertEquals(MpResult.QUESTIONABLE, output.getResult());
 
         //M13 Abstract multiple primaries** when a neoplasm is originally diagnosed as acute AND reverts to a chronic neoplasm after treatment
-        i1.setTxStatus("1"); //treatment was done for acute
+        i1.setTxStatus("1");
         output = _utils.computePrimaries(i1, i2);
-        Assert.assertEquals(13, output.getAppliedRules().size());
-        Assert.assertEquals(MphUtils.MpResult.MULTIPLE_PRIMARIES, output.getResult());
+        Assert.assertEquals(12, output.getAppliedRules().size());
+        Assert.assertEquals(MpResult.QUESTIONABLE, output.getResult());
 
         //M14 Abstract a single primary* when post-transplant lymphoproliferative disorder is diagnosed simultaneously with any B-cell lymphoma, T-cell
         //lymphoma, Hodgkin lymphoma or plasmacytoma/myeloma
@@ -2104,7 +2085,7 @@ public class MphUtilsTest {
         output = _utils.computePrimaries(i1, i2);
         Assert.assertEquals(14, output.getAppliedRules().size());
         Assert.assertEquals(MphUtils.MpResult.SINGLE_PRIMARY, output.getResult());
-        i2.setHistologyIcdO3("9596"); //Hodgkin
+        i2.setHistologyIcdO3("9653"); //Hodgkin
         output = _utils.computePrimaries(i1, i2);
         Assert.assertEquals(14, output.getAppliedRules().size());
         Assert.assertEquals(MphUtils.MpResult.SINGLE_PRIMARY, output.getResult());